Expose withTimeout helper from utils
It's not specific to FSM usage at all.
